// src/data/enum/action.enum.ts
export enum ActionEnum {
  THEME_UPDATE = 'THEME_UPDATE',
  //
  WORDLIST_APPEND = 'WORDLIST_APPEND',
  WORDLIST_FOUND_WORD = 'WORDLIST_FOUND_WORD',
  WORDLIST_CANCEL_WORD = 'WORDLIST_CANCEL_WORD',
  WORDLIST_CONFIRM_WORDS = 'WORDLIST_CONFIRM_WORDS',
  //
  SCORE_ADD = 'SCORE_ADD',
  //
  TIMER_START = 'TIMER_START',
  TIMER_TICK = 'TIMER_TICK',
  TIMER_STOP = 'TIMER_STOP',
  //
  APP_START_GAME = 'APP_START_GAME',
  APP_END_GAME = 'APP_END_GAME',
  APP_RESET_GAME = 'APP_RESET_GAME',
  APP_START_PHASE = 'APP_START_PHASE',
  APP_END_PHASE = 'APP_END_PHASE',
  APP_START_ROUND = 'APP_START_ROUND',
  APP_END_ROUND = 'APP_END_ROUND',
  APP_CONFIRM_ROUND = 'APP_CONFIRM_ROUND',
}